requirements.txt
Pillow
requests
tzlocal
pytz
python-dotenv